+++ js/InitData.js
@@ -1,0 +1,18 @@
+export const Init = {
+  VALUES_QUANTITY: 25,
+  MIN_RANGE_ID: 1,
+  MAX_RANGE_ID: 25,
+  MIN_RANGE_LIKES: 15,
+  MAX_RANGE_LIKES: 200,
+  MIN_RANGE_COMMENTS: 0,
+  MAX_RANGE_COMMENTS: 30,
+  MIN_RANGE_AVATAR: 1,
+  MAX_RANGE_AVATAR: 6,
+  NAMES: ['Николай', 'Татьяна', 'Владимир', 'Виктория', 'Матвей', 'Екатерина', 'Павел', 'Людмила'],
+  MESSAGES: `Всё отлично!
+В целом всё неплохо. Но не всё.
+Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.
+Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.
+Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.
+Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!`,
+};
